fix: Update docker install variable default value
The docker_install variable in the JupyterHub playbook YAML file has
been updated to have a default value of true. This change ensures that
the variable is set correctly when no value is provided.

feat: Add list_files to zimbra installation tasks

The list_files option has been added to the Zimbra installation tasks
for CentOS 8, CentOS 7, Ubuntu 20.04, and Ubuntu 18.04. This option
allows the tasks to list the files in the Zimbra installation archive.

refactor: Update chdir path in zimbra installer tasks

The chdir path in the Zimbra installer tasks for CentOS 8, CentOS 7,
Ubuntu 20.04, and Ubuntu 18.04 has been updated to use the path of the
first file in the zimbra_zip variable. This change ensures that the
installer is run on the correct Zimbra version.