feat: Add support for configuration mapping in Apache NiFi Docker deployment
The code changes introduce support for mapping the configuration files
in the Apache NiFi Docker deployment. If the `conf_mapping` flag is set
to `true`, the configuration files will be copied from the containers to
the `./conf` directory. Additionally, directory permissions for `./conf`
are set to `777`. This enables the restarting of services using the
updated configuration. The play will wait until the Apache NiFi URL is
reachable before proceeding.