feat: Enable LDAP authentication in Penpot configuration
This commit enables LDAP authentication in the Penpot configuration. The
code changes include the addition of LDAP-related fields in the
platform-configuration-fields.json file and the introduction of
LDAP-specific variables in the .env.auth.j2 file. Additionally, a new
ldap.yaml task was created to handle the LDAP configuration process.

Enabling LDAP authentication provides users with the ability to log in
to Penpot using their LDAP credentials, enhancing security and
streamlining authentication for organizations utilizing LDAP services.